requirements.txt
Django==2.2.27
djangorestframework==3.13.1
djoser==2.1.0
Pillow==9.0.1
pytz==2021.3
sqlparse==0.4.2
